requirements.txt
gradio
flask
python-dotenv
gunicorn
requests
google-auth
google-auth-oauthlib
google-api-python-client
huggingface_hub
dnspython
proxy.py